+++ flink-python/pyflink/fn_execution/beam/beam_operations_fast.pxd
@@ -20,13 +20,13 @@
 from apache_beam.coders.coder_impl cimport StreamCoderImpl
 from apache_beam.runners.worker.operations cimport Operation
 
-from pyflink.fn_execution.coder_impl_fast cimport BaseCoderImpl
+from pyflink.fn_execution.coder_impl_fast cimport LengthPrefixBaseCoderImpl
 
 cdef class FunctionOperation(Operation):
     cdef Operation consumer
     cdef bint _is_python_coder
     cdef StreamCoderImpl _value_coder_impl
-    cdef BaseCoderImpl _output_coder
+    cdef LengthPrefixBaseCoderImpl _output_coder
     cdef object process_element
     cdef object operation
     cdef object operation_cls
